requirements.txt
aiosqlite==0.20.0
alembic==1.16.4
annotated-types==0.7.0
asyncpg==0.30.0
bcrypt==4.0.1
passlib==1.7.4
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic-settings==2.5.2
pydantic_core==2.33.2
PyJWT==2.10.1
SQLAlchemy==2.0.41
sqlmodel==0.0.24
fastapi
pyqt6
websockets